+++ packages/cli/src/commands/codegen.ts
@@ -1,5 +1,4 @@
-import { CLIOptions, Inquirerer } from 'inquirerer';
-import { ParsedArgs } from 'minimist';
+import { CLIOptions, Inquirerer, ParsedArgs } from 'inquirerer';
 import * as fs from 'node:fs';
 import * as path from 'node:path';
 import { buildSchemaSDL } from '@constructive-io/graphql-server';