+++ pgpm/cli/src/commands/docker.ts
@@ -1,6 +1,5 @@
 import { spawn } from 'child_process';
-import { cliExitWithError, extractFirst } from '@inquirerer/utils';
-import { CLIOptions, Inquirerer } from 'inquirerer';
+import { CLIOptions, Inquirerer, cliExitWithError, extractFirst } from 'inquirerer';
 
 const dockerUsageText = `
 Docker Command: