+++ pgpm/cli/src/commands/tag.ts
@@ -1,8 +1,7 @@
-import { extractFirst } from '@inquirerer/utils';
 import { PgpmPackage } from '@pgpmjs/core';
 import { Logger } from '@pgpmjs/logger';
 import { errors } from '@pgpmjs/types';
-import { CLIOptions, Inquirerer, Question } from 'inquirerer';
+import { CLIOptions, Inquirerer, Question, extractFirst } from 'inquirerer';
 import * as path from 'path';
 import { selectPackage } from '../utils/module-utils';
 import { resolvePackageAlias } from '../utils/package-alias';