+++ packages/cli/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.10.3](https://github.com/constructive-io/constructive/compare/@constructive-io/cli@5.10.2...@constructive-io/cli@5.10.3) (2026-01-21)
+
+**Note:** Version bump only for package @constructive-io/cli
+
+## [5.10.2](https://github.com/constructive-io/constructive/compare/@constructive-io/cli@5.10.1...@constructive-io/cli@5.10.2) (2026-01-20)
+
+**Note:** Version bump only for package @constructive-io/cli
+
+## [5.10.1](https://github.com/constructive-io/constructive/compare/@constructive-io/cli@5.10.0...@constructive-io/cli@5.10.1) (2026-01-20)
+
+**Note:** Version bump only for package @constructive-io/cli
+
+# [5.10.0](https://github.com/constructive-io/constructive/compare/@constructive-io/cli@5.9.3...@constructive-io/cli@5.10.0) (2026-01-20)
+
+**Note:** Version bump only for package @constructive-io/cli
+
+## [5.9.3](https://github.com/constructive-io/constructive/compare/@constructive-io/cli@5.9.2...@constructive-io/cli@5.9.3) (2026-01-20)
+
+**Note:** Version bump only for package @constructive-io/cli
+
+## [5.9.2](https://github.com/constructive-io/constructive/compare/@constructive-io/cli@5.9.1...@constructive-io/cli@5.9.2) (2026-01-19)
+
+**Note:** Version bump only for package @constructive-io/cli
+
 ## [5.9.1](https://github.com/constructive-io/constructive/compare/@constructive-io/cli@5.9.0...@constructive-io/cli@5.9.1) (2026-01-18)
 
 **Note:** Version bump only for package @constructive-io/cli